+++ ui/src/main/java/io/snabble/sdk/ui/utils/VerticalSwipeRefreshLayout.java
@@ -1,0 +1,40 @@
+package io.snabble.sdk.ui.utils;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.view.MotionEvent;
+import android.view.ViewConfiguration;
+
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+
+public class VerticalSwipeRefreshLayout extends SwipeRefreshLayout {
+    private int touchSlop;
+    private float prevX;
+    private boolean declined;
+
+    public VerticalSwipeRefreshLayout(Context context, AttributeSet attrs) {
+        super(context, attrs);
+        touchSlop = ViewConfiguration.get(context).getScaledTouchSlop();
+    }
+
+    @Override
+    public boolean onInterceptTouchEvent(MotionEvent event) {
+        switch (event.getAction()) {
+            case MotionEvent.ACTION_DOWN:
+                prevX = MotionEvent.obtain(event).getX();
+                declined = false;
+                break;
+
+            case MotionEvent.ACTION_MOVE:
+                final float eventX = event.getX();
+                float xDiff = Math.abs(eventX - prevX);
+                if (declined || xDiff > touchSlop) {
+                    declined = true;
+                    return false;
+                }
+                break;
+        }
+
+        return super.onInterceptTouchEvent(event);
+    }
+}